Dodate addEventListener metode i HTML "očišćen" od JS-a
