+++ CShop v3/js/myScript.js
@@ -1,337 +1,19 @@
-// var notebookData = null;
+import notebookData from './modules/myData.js';
 
-// console.log(notebookData);
-// // JavaScript is a "hoisting" programming language
-
-document.body.onload = function(){
-    filterData();
-};
+window.addEventListener("load", filterData);
 
 var notebookDataObject = document.getElementById("notebookData");
 
-var notebookData = [
-    {
-        brandName: "Acer",
-        modelName: "Aspire A515-45-R0QQ",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana",
-        PriceOld: 74.999,
-        PriceNew: 65.999,
-        ThumbImage: "imgs/ACER Aspire A515-45-R0QQ.png",
-    },
-    {
-        brandName: "Acer",
-        modelName: "Predator Helios 300",
-        ramMemory: "32GB",
-        SSD: "1TB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce RTX 3070 Ti",
-        PriceOld: 399.999,
-        PriceNew: 299.989,
-        ThumbImage: "imgs/ACER Predator Helios 300.png"
-    },
-    {
-        brandName: "Acer",
-        modelName: "Aspire 3",
-        ramMemory: "4GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "TN",
-        Graphics: "Integrisana",
-        PriceOld: 55.999,
-        PriceNew: 54.999,
-        ThumbImage: "imgs/ACER Aspire 3 NX.HS5EX.01N.png"
-    },
-    {
-        brandName: "Acer",
-        modelName: "Nitro 5",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 17.3,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce GTX 1650",
-        PriceOld: 104.999,
-        PriceNew: 99.999,
-        ThumbImage: "imgs/ACER Nitro 5 AN517-52.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "Legion 5 17ACH6H",
-        ramMemory: "16GB",
-        SSD: "1TB",
-        ScreenDiagonal: 17.3,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce RTX 3070",
-        PriceOld: 234.999,
-        PriceNew: 234.999,
-        ThumbImage: "imgs/LENOVO Legion 5 17ACH6H.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "Legion 5 15IMH6",
-        ramMemory: "8GB",
-        SSD: "512GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce RTX 3050",
-        PriceOld: 124.999,
-        PriceNew: 124.999,
-        ThumbImage: "imgs/LENOVO Legion 5 15IMH6.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "Legion 5 Pro 16ARH7H",
-        ramMemory: "16GB",
-        SSD: "1TB",
-        ScreenDiagonal: 16,
-        ScreenResolution: "2560 x 1600",
-        ScreenType: "IPS",
-        Graphics: "GeForce RTX 3060",
-        PriceOld: 239.999,
-        PriceNew: 209.999,
-        ThumbImage: "imgs/LENOVO Legion 5 Pro 16ARH7H.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "IdeaPad 3 15ABA7",
-        ramMemory: "16GB",
-        SSD: "512GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana",
-        PriceOld: 111.109,
-        PriceNew: 83.329,
-        ThumbImage: "imgs/LENOVO IdeaPad 3 15ABA7.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "IdeaPad 3 15ITL6",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "TN",
-        Graphics: "Integrisana",
-        PriceOld: 77.779,
-        PriceNew: 58.329,
-        ThumbImage: "imgs/LENOVO IdeaPad 3 15ITL6.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "IdeaPad 5 15ALC05",
-        ramMemory: "16GB",
-        SSD: "512GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana",
-        PriceOld: 105.559,
-        PriceNew: 79.159,
-        ThumbImage: "imgs/LENOVO IdeaPad 5 15ALC05.png"
-    },
-    {
-        brandName: "Lenovo",
-        modelName: "V15 G2 ITLi5",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "TN",
-        Graphics: "Integrisana",
-        PriceOld: 84.449,
-        PriceNew: 63.329,
-        ThumbImage: "imgs/Lenovo V15 G2 ITLi5.png"
-    },
-    {
-        brandName: "Apple",
-        modelName: "MacBook Air 13 Retina Space Grey MGN63ZEA",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 13.3,
-        ScreenResolution: "2560 x 1600",
-        ScreenType: "Retina",
-        Graphics: "Integrisana Apple M1",
-        PriceOld: 164.899,
-        PriceNew: 143.900,
-        ThumbImage: "imgs/APPLE MacBook Air 13 Retina Space Grey MGN63ZEA.png"
-    },
-    {
-        brandName: "Apple",
-        modelName: "MacBook Air 13 Retina Gold MGND3ZEA",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 13.3,
-        ScreenResolution: "2560 x 1600",
-        ScreenType: "Retina",
-        Graphics: "Integrisana Apple M1",
-        PriceOld: 164.899,
-        PriceNew: 143.900,
-        ThumbImage: "imgs/APPLE MacBook Air 13 Retina Gold MGND3ZEA.png"
-    },
-    {
-        brandName: "Apple",
-        modelName: "MacBook Air 13 Z124000KK",
-        ramMemory: "16GB",
-        SSD: "256GB",
-        ScreenDiagonal: 13.3,
-        ScreenResolution: "2560 x 1600",
-        ScreenType: "Retina",
-        Graphics: "Integrisana Apple M1",
-        PriceOld: 189.999,
-        PriceNew: 189.999,
-        ThumbImage: "imgs/MacBook Air 13 Z124000KK.png"
-    },
-    {
-        brandName: "Apple",
-        modelName: "MacBook Air M2 Midnight 24GB/1TB ",
-        ramMemory: "24GB",
-        SSD: "1TB",
-        ScreenDiagonal: 13.3,
-        ScreenResolution: "2560 x 1600",
-        ScreenType: "Retina",
-        Graphics: "Integrisana Apple M2",
-        PriceOld: 329.999,
-        PriceNew: 329.999,
-        ThumbImage: "imgs/APPLE MacBook Air M2 Midnight 24GB-1TB.png"
-    },
-    {
-        brandName: "Apple",
-        modelName: "MacBook Pro 16 M1 Max 2021 Space Grey MK1A3/BTO/64GB/8TB",
-        ramMemory: "64GB",
-        SSD: "2568TB",
-        ScreenDiagonal: 16.2,
-        ScreenResolution: "3456 x 2234",
-        ScreenType: "Liquid Retin XDR",
-        Graphics: "Integrisana Apple M1",
-        PriceOld: 899.999,
-        PriceNew: 899.999,
-        ThumbImage: "imgs/APPLE MacBook Pro 16 M1 Max 2021 Space Grey.png"
-    },
-    {
-        brandName: "Dell",
-        modelName: "Inspiron 3502",
-        ramMemory: "4GB",
-        SSD: "2568TB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "MVA",
-        Graphics: "Integrisana UHD 605",
-        PriceOld: 45.999,
-        PriceNew: 43.999,
-        ThumbImage: "imgs/DELL Inspiron 3502.png"
-    },
-    {
-        brandName: "Dell",
-        modelName: "Vostro 3510",
-        ramMemory: "8GB",
-        SSD: "2568TB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "MVA",
-        Graphics: "Integrisana UHD 605",
-        PriceOld: 105.559,
-        PriceNew: 79.159,
-        ThumbImage: "imgs/DELL Vostro 3510.png"
-    },
-    {
-        brandName: "Dell",
-        modelName: "G15 5520",
-        ramMemory: "16GB",
-        SSD: "5128TB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "TN",
-        Graphics: "GeForce RTX 3050 Ti",
-        PriceOld: 255.559,
-        PriceNew: 191.659,
-        ThumbImage: "imgs/DELL G15 5520.png"
-    },
-    {
-        brandName: "HP",
-        modelName: "Omen 16-k0007na",
-        ramMemory: "16GB",
-        SSD: "512GB",
-        ScreenDiagonal: 16.1,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce RTX 3060",
-        PriceOld: 249.999,
-        PriceNew: 249.999,
-        ThumbImage: "imgs/HP Omen 16-k0007na.png"
-    },
-    {
-        brandName: "HP",
-        modelName: "250 G8 - 3V5L7EA",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana UHD Graphics",
-        PriceOld: 66.669,
-        PriceNew: 49.999,
-        ThumbImage: "imgs/HP 250 G8 - 3V5L7EA.png"
-    },
-    {
-        brandName: "HP",
-        modelName: "ProBook 440 G8",
-        ramMemory: "8GB",
-        SSD: "512GB",
-        ScreenDiagonal: 14,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana Iris Xe Graphics",
-        PriceOld: 111.109,
-        PriceNew: 83.329,
-        ThumbImage: "imgs/HP ProBook 440 G8 - 2X7U7EA.png"
-    },
-    {
-        brandName: "HP",
-        modelName: "Victus 16-d1003nm - 65C75EA",
-        ramMemory: "8GB",
-        SSD: "512GB",
-        ScreenDiagonal: 16.1,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "GeForce GTX 1650",
-        PriceOld: 155.559,
-        PriceNew: 116.659,
-        ThumbImage: "imgs/HP Victus 16-d1003nm - 65C75EA.png"
-    },
-    {
-        brandName: "Fujitsu",
-        modelName: "Lifebook A3510",
-        ramMemory: "8GB",
-        SSD: "256GB",
-        ScreenDiagonal: 15.6,
-        ScreenResolution: "1920 x 1080",
-        ScreenType: "IPS",
-        Graphics: "Integrisana UHD G1",
-        PriceOld: 61.999,
-        PriceNew: 57.999,
-        ThumbImage: "imgs/FUJITSU Lifebook A3510 FPC04967BP.png"
-    }
-];
-
-
 var onDiscountChecked = document.getElementById('onDiscount');
+onDiscountChecked.addEventListener('click', filterData);
+
 var notOnDiscountChecked = document.getElementById('notOnDiscount');
+notOnDiscountChecked.addEventListener('click', filterData);
 
 function filterData() {
-    tmpString = "";
+    var tmpString = "";
 
-    for (i = 0; i < notebookData.length; i++) {
+    for (let i = 0; i < notebookData.length; i++) {
         if (onDiscountChecked.checked & notebookData[i].PriceOld != notebookData[i].PriceNew) {
             tmpString += createHTMLData(notebookData[i]);
         }
@@ -349,7 +31,6 @@ function filterData() {
         notebookDataObject.innerHTML = tmpString;
     }
 }
-
 
 function createHTMLData(notebookItem) {
     return `<div class="col-12 col-md-4 col-lg-3 mb-5">
