+++ CShop v3/notebook-cards.html
@@ -51,11 +51,11 @@
 
     <form class="mb-5">
       <div class="custom-control custom-checkbox">
-        <input type="checkbox" class="custom-control-input" id="onDiscount" onclick="filterData()" checked>
+        <input type="checkbox" class="custom-control-input" id="onDiscount" checked>
         <label class="custom-control-label" for="onDiscount">Na popustu!</label>
       </div>
       <div class="custom-control custom-checkbox">
-        <input type="checkbox" class="custom-control-input" id="notOnDiscount" onclick="filterData()" checked>
+        <input type="checkbox" class="custom-control-input" id="notOnDiscount" checked>
         <label class="custom-control-label" for="notOnDiscount">Nije na popustu!</label>
       </div>
     </form>
@@ -84,7 +84,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.min.js" integrity="sha384-+sLIOodYLS7CIrQpBjl+C7nPvqq+FbNUBDunl/OZv93DB7Ln/533i8e/mZXLi/P+" crossorigin="anonymous"></script>
     -->
 
-  <script src="./js/myScript.js"></script>
+  <script type="module" src="./js/myScript.js"></script>
 </body>
 
 </html>
